seperate auto stop from task logic
