+++ src/aind_behavior_dynamic_foraging/CurriculumManager/curriculums/uncoupled_baiting_2p3.py
@@ -18,7 +18,6 @@ from aind_behavior_dynamic_foraging.DataSchemas.task_logic import (
     AutoWater,
     InterTrialInterval,
     ResponseTime,
-    AutoStop,
     AutoBlock,
     RewardSize,
     Warmup
@@ -40,7 +39,6 @@ s_stage_1_warmup = Stage(name="stage_1_warmup", task=AindDynamicForagingTaskLogi
         reward_size=RewardSize(right_value_volume=4.0, left_value_volume=4.0),
         auto_water=AutoWater(auto_water_type=AutoWaterMode.NATURAL, unrewarded=3, ignored=3, multiplier=0.5),
         auto_block=AutoBlock(advanced_block_auto="now", switch_thr=0.5, points_in_a_row=5),
-        auto_stop=AutoStop(ignore_win=20000, ignore_ratio_threshold=1, max_trial=1000, max_time=75),
         response_time=ResponseTime(response_time=5, reward_consume_time=1),
         uncoupled_reward=None
     )
@@ -56,7 +54,6 @@ s_stage_1 = Stage(name="stage_1", task=AindDynamicForagingTaskLogic(
         reward_size=RewardSize(right_value_volume=2.0, left_value_volume=2.0),
         auto_water=AutoWater(auto_water_type=AutoWaterMode.NATURAL, unrewarded=5, ignored=5, multiplier=0.5),
         auto_block=AutoBlock(advanced_block_auto="now", switch_thr=0.5, points_in_a_row=5),
-        auto_stop=AutoStop(ignore_win=20000, ignore_ratio_threshold=1, max_trial=1000, max_time=75),
         response_time=ResponseTime(response_time=5, reward_consume_time=1),
         uncoupled_reward=None
     )
@@ -72,7 +69,6 @@ s_stage_2 = Stage(name="stage_2", task=AindDynamicForagingTaskLogic(
         reward_size=RewardSize(right_value_volume=2.0, left_value_volume=2.0),
         auto_water=AutoWater(auto_water_type=AutoWaterMode.NATURAL, unrewarded=7, ignored=7, multiplier=0.5),
         auto_block=AutoBlock(advanced_block_auto="now", switch_thr=0.5, points_in_a_row=5),
-        auto_stop=AutoStop(ignore_win=30, ignore_ratio_threshold=0.83, max_trial=1000, max_time=75),
         response_time=ResponseTime(response_time=3, reward_consume_time=1),
         uncoupled_reward=None
     )
@@ -87,7 +83,6 @@ s_stage_3 = Stage(name="stage_3", task=AindDynamicForagingTaskLogic(
         reward_delay=0.0,
         reward_size=RewardSize(right_value_volume=2.0, left_value_volume=2.0),
         auto_water=AutoWater(auto_water_type=AutoWaterMode.NATURAL, unrewarded=10, ignored=10, multiplier=0.5),
-        auto_stop=AutoStop(ignore_win=30, ignore_ratio_threshold=0.83, max_trial=1000, max_time=75),
         response_time=ResponseTime(response_time=2, reward_consume_time=1),
         uncoupled_reward=[0.1, 0.4, 0.7]
     )
@@ -101,7 +96,6 @@ s_final = Stage(name="final", task=AindDynamicForagingTaskLogic(
         delay_period=DelayPeriod(min=1, max=1, beta=0),
         reward_delay=0.0,
         reward_size=RewardSize(right_value_volume=2.0, left_value_volume=2.0),
-        auto_stop=AutoStop(ignore_win=30, ignore_ratio_threshold=0.83, max_trial=1000, max_time=75),
         response_time=ResponseTime(response_time=1, reward_consume_time=3),
         uncoupled_reward=[0.1, 0.4, 0.7]
     )
@@ -115,7 +109,6 @@ s_graduated = Stage(name="graduated", task=AindDynamicForagingTaskLogic(
         delay_period=DelayPeriod(min=1, max=1, beta=0),
         reward_delay=0.0,
         reward_size=RewardSize(right_value_volume=2.0, left_value_volume=2.0),
-        auto_stop=AutoStop(ignore_win=30, ignore_ratio_threshold=0.83, max_trial=1000, max_time=75),
         response_time=ResponseTime(response_time=1, reward_consume_time=3),
         uncoupled_reward=[0.1, 0.4, 0.7]
     )
