+++ src/aind_behavior_dynamic_foraging/CurriculumManager/curriculums/uncoupled_no_baiting_2p3.py
@@ -15,7 +15,6 @@ from aind_behavior_dynamic_foraging.DataSchemas.task_logic import (
     AutoWater,
     InterTrialInterval,
     ResponseTime,
-    AutoStop,
     AutoBlock,
     RewardSize,
     Warmup
@@ -76,13 +75,6 @@ s_stage_1_warmup = Stage(
                 switch_thr=0.5,
                 points_in_a_row=5
             ),
-            auto_stop=AutoStop(
-                ignore_win=20000,
-                ignore_ratio_threshold=1,
-                max_trial=1000,
-                max_time=75,
-                min_time=30
-            ),
             response_time=ResponseTime(
                 response_time=5,
                 reward_consume_time=1
@@ -133,13 +125,6 @@ s_stage_1 = Stage(
                 advanced_block_auto="now",
                 switch_thr=0.5,
                 points_in_a_row=5
-            ),
-            auto_stop=AutoStop(
-                ignore_win=20000,
-                ignore_ratio_threshold=1,
-                max_trial=1000,
-                max_time=75,
-                min_time=30
             ),
             response_time=ResponseTime(
                 response_time=5,
@@ -192,13 +177,6 @@ s_stage_2 = Stage(
                 switch_thr=0.5,
                 points_in_a_row=5
             ),
-            auto_stop=AutoStop(
-                ignore_win=30,
-                ignore_ratio_threshold=0.83,
-                max_trial=1000,
-                max_time=75,
-                min_time=30
-            ),
             response_time=ResponseTime(
                 response_time=1.5,
                 reward_consume_time=1
@@ -250,13 +228,6 @@ s_stage_3 = Stage(
                 switch_thr=0.5,
                 points_in_a_row=5
             ),
-            auto_stop=AutoStop(
-                ignore_win=30,
-                ignore_ratio_threshold=0.83,
-                max_trial=1000,
-                max_time=75,
-                min_time=30
-            ),
             response_time=ResponseTime(
                 response_time=1.5,
                 reward_consume_time=1
@@ -303,13 +274,6 @@ s_stage_4 = Stage(
                 unrewarded=10,
                 ignored=10,
             ),
-            auto_stop=AutoStop(
-                ignore_win=30,
-                ignore_ratio_threshold=0.83,
-                max_trial=1000,
-                max_time=75,
-                min_time=30
-            ),
             response_time=ResponseTime(
                 response_time=1.5,
                 reward_consume_time=1
@@ -350,13 +314,6 @@ s_final = Stage(
                 right_value_volume=2.0,
                 left_value_volume=2.0
             ),
-            auto_stop=AutoStop(
-                ignore_win=30,
-                ignore_ratio_threshold=0.83,
-                max_trial=1000,
-                max_time=75,
-                min_time=30
-            ),
             response_time=ResponseTime(
                 response_time=1.5,
                 reward_consume_time=1
@@ -395,13 +352,6 @@ s_graduated = Stage(
             reward_size=RewardSize(
                 right_value_volume=2.0,
                 left_value_volume=2.0
-            ),
-            auto_stop=AutoStop(
-                ignore_win=30,
-                ignore_ratio_threshold=0.83,
-                max_trial=1000,
-                max_time=75,
-                min_time=30
             ),
             response_time=ResponseTime(
                 response_time=1.5,
